Use proxy wiring for event hasher
