+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/consensus/TestIntake.java
@@ -24,6 +24,7 @@ import static org.mockito.Mockito.mock;
 import com.swirlds.base.time.Time;
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.platform.NodeId;
+import com.swirlds.common.wiring.component.ComponentWiring;
 import com.swirlds.common.wiring.model.WiringModel;
 import com.swirlds.common.wiring.schedulers.TaskScheduler;
 import com.swirlds.common.wiring.schedulers.builders.TaskSchedulerType;
@@ -33,6 +34,7 @@ import com.swirlds.platform.components.ConsensusEngine;
 import com.swirlds.platform.consensus.ConsensusSnapshot;
 import com.swirlds.platform.consensus.NonAncientEventWindow;
 import com.swirlds.platform.event.GossipEvent;
+import com.swirlds.platform.event.hashing.DefaultEventHasher;
 import com.swirlds.platform.event.hashing.EventHasher;
 import com.swirlds.platform.event.linking.InOrderLinker;
 import com.swirlds.platform.event.orphan.OrphanBuffer;
@@ -49,7 +51,6 @@ import com.swirlds.platform.test.fixtures.event.IndexedEvent;
 import com.swirlds.platform.wiring.ConsensusEngineWiring;
 import com.swirlds.platform.wiring.InOrderLinkerWiring;
 import com.swirlds.platform.wiring.OrphanBufferWiring;
-import com.swirlds.platform.wiring.components.EventHasherWiring;
 import com.swirlds.platform.wiring.components.EventWindowManagerWiring;
 import com.swirlds.platform.wiring.components.PostHashCollectorWiring;
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -66,7 +67,7 @@ public class TestIntake implements LoadableFromSignedState {
     private final Shadowgraph shadowGraph;
     private final ConsensusOutput output;
 
-    private final EventHasherWiring hasherWiring;
+    private final ComponentWiring<EventHasher, GossipEvent> hasherWiring;
     private final OrphanBufferWiring orphanBufferWiring;
     private final InOrderLinkerWiring linkerWiring;
     private final ConsensusEngineWiring consensusEngineWiring;
@@ -88,8 +89,8 @@ public class TestIntake implements LoadableFromSignedState {
 
         model = WiringModel.create(platformContext, time, mock(ForkJoinPool.class));
 
-        final EventHasher eventHasher = new EventHasher(platformContext);
-        hasherWiring = EventHasherWiring.create(directScheduler("eventHasher"));
+        hasherWiring = new ComponentWiring<>(model, EventHasher.class, directScheduler("eventHasher"));
+        final EventHasher eventHasher = new DefaultEventHasher(platformContext);
         hasherWiring.bind(eventHasher);
 
         final PostHashCollectorWiring postHashCollectorWiring =
@@ -112,7 +113,7 @@ public class TestIntake implements LoadableFromSignedState {
 
         final EventWindowManagerWiring eventWindowManagerWiring = EventWindowManagerWiring.create(model);
 
-        hasherWiring.eventOutput().solderTo(postHashCollectorWiring.eventInput());
+        hasherWiring.getOutputWire().solderTo(postHashCollectorWiring.eventInput());
         postHashCollectorWiring.eventOutput().solderTo(orphanBufferWiring.eventInput());
         orphanBufferWiring.eventOutput().solderTo(linkerWiring.eventInput());
         linkerWiring.eventOutput().solderTo("shadowgraph", "addEvent", shadowGraph::addEvent);
@@ -140,7 +141,7 @@ public class TestIntake implements LoadableFromSignedState {
      * @param event the event to add
      */
     public void addEvent(@NonNull final GossipEvent event) {
-        hasherWiring.eventInput().put(event);
+        hasherWiring.getInputWire(EventHasher::hashEvent).put(event);
     }
 
     /**
